==> code/notebooks/1. Processing - RML Data.ipynb <==
# Processing and Writing Radiant ML Data to CSV

This notebook focuses on extracting AgriFieldNet Competition field data from GeoTiffs and saving the results to CSV. Much of the code to perform these tasks originates with Radiant Earths' [starter notebook](https://github.com/radiantearth/agrifieldnet_india_competition/blob/main/Starter%20notebook.ipynb), but have been heavily modified to meet the specific goals of the author.

*Note: The related sibling notebook, 'Processing - RML Data (Geospatial)' goes in greater depth and breaks up and recreates scaled per-field GeoTiffs for each band. However, the results from the Geospatial notebook were not ultimately used by this project.*

## Setup

Create the libraries + import config

import json
import pandas as pd
import matplotlib.pyplot as plt

import sys
sys.path.append('..')

from utils.file_ops import *
from utils.data_prep import *
from utils.vegetation_indices import *

# get config
CONFIG = read_yaml('../../conf.yaml')

Setting up vars from config

data_dir = '../../data/'

Full_bands = CONFIG['BANDS']
collection_id = CONFIG['COLLECTION_ID']
assets = CONFIG['ASSETS']
processed_collection_id = CONFIG['PROCESSED_COLLECTION_ID']

crops = [c['id'] for c in CONFIG['CROPS']]
crops_names = [c['name'] for c in CONFIG['CROPS']]

selected_bands = [b['id'] for b in CONFIG['BANDS']]
source_collection = f'{collection_id}_source'
train_label_collection = f'{collection_id}_labels_train'
test_label_collection = f'{collection_id}_labels_test'

src_tile_size = (CONFIG['SRC_TILE_WIDTH'], CONFIG['SRC_TILE_HEIGHT'])

## Download the data if needed

# requires MLHUB_API_KEY ENV var to be set

def fetch_rml_data(collection_id, assets, selected_bands):
    target_dir = data_dir + collection_id

    if exists(target_dir):
        print(f'the directory {collection_id} already exists.')
    else:
        try:
            dataset = Dataset.fetch(collection_id)

            my_filter = dict(
                ref_agrifieldnet_competition_v1_labels_train=assets,
                ref_agrifieldnet_competition_v1_labels_test=[assets[0]],
                ref_agrifieldnet_competition_v1_source=selected_bands 
            )

            dataset.download(output_dir=target_dir, collection_filter=my_filter)
            print(f'{collection_id} successfully downloaded!')
        except:
            print(f'fetch_rml_data failed to download {collection_id} data!')


fetch_rml_data(collection_id, assets, selected_bands)

## Prepare Train data

def get_field_and_label_paths(collection_id, label_collection, bin='train'):
    with open (f'{data_dir}{collection_id}/{label_collection}/collection.json') as f:
        json_data = json.load(f)
        
    folder_ids = [i['href'].split('_')[-1].split('.')[0] for i in json_data['links'][4:]]
    field_paths = [f'{data_dir}{collection_id}/{label_collection}/{label_collection}_{i}/field_ids.tif' for i in folder_ids]
    if bin == 'train':
        label_paths = [f'{data_dir}{collection_id}/{label_collection}/{label_collection}_{i}/raster_labels.tif' for i in folder_ids]    
    else:
        label_paths = []
    return folder_ids, field_paths, label_paths

train_folder_ids, train_field_paths, train_label_paths = get_field_and_label_paths(collection_id, train_label_collection)
test_folder_ids, test_field_paths, test_label_paths = get_field_and_label_paths(collection_id, test_label_collection, 'test')

### Train

#create dataset for folder_ids and field_paths

competition_train_data = pd.DataFrame(train_folder_ids, columns=['unique_folder_id'])
competition_train_data['field_paths'] = train_field_paths
write_csv_from_df(
    competition_train_data, 
    f'{data_dir}interstitial_data/folder_ids_and_field_paths_TRAIN'
)

print(competition_train_data.shape)
competition_train_data.head()

### Test

#create dataset for folder_ids and field_paths

competition_test_data = pd.DataFrame(test_folder_ids, columns=['unique_folder_id'])
competition_test_data['field_paths'] = test_field_paths
write_csv_from_df(
    competition_test_data, 
    f'{data_dir}interstitial_data/folder_ids_and_field_paths_TEST'
)

print(competition_test_data.shape)
competition_test_data.head()

## PREPROCESS FIELDS AND CROPS IN TILES FOR TRAININIG

### Train

field_crop_data = get_df_from_csv_if_exists(
    f'{data_dir}labels_TRAIN.csv',
    lambda: field_crop_extractor(train_folder_ids, collection_id, train_label_collection)
)

write_csv_from_df(field_crop_data, f'{data_dir}labels_TRAIN')
print(field_crop_data.shape)
field_crop_data.head()

img_sh = 256
n_obs = 1  #imagery per chip(no time series)

train_data = get_df_from_csv_if_exists(
    f'{data_dir}/pixel_data_TRAIN.csv',
    lambda: pixel_data_extractor_1(competition_train_data, source_collection, 
                              collection_id, selected_bands, n_obs, img_sh)
)

write_csv_from_df(train_data, f'{data_dir}/pixel_data_TRAIN')
print(train_data.shape)
train_data.head()

### Test

test_data = get_df_from_csv_if_exists(
    f'{data_dir}pixel_data_TEST.csv',
    lambda: pixel_data_extractor_1(competition_test_data, source_collection, 
                              collection_id, selected_bands, n_obs, img_sh)
)

write_csv_from_df(test_data, f'{data_dir}pixel_data_TEST')
print(test_data.shape)
test_data.head()

## Adding Vegetation Indices

Vegetation Indices (VIs) combine surface reflectance at two or more wavelengths and highlight a given property of vegetation. They are derived using the reflectance properties of vegetation. This project utilizes the following VIs

- NDVI
- ARVI
- SAVI
- SRI
- RENDVI
- ARI
- SAVI
- MSI
- MCARI
- MARI
- EVI2
- NDMI
- NDWI

add_vegetation_indices(train_data, selected_bands)
add_vegetation_indices(test_data, selected_bands)

## Aggregate the data

train_data.columns

### Train

train_data_agg = get_df_from_csv_if_exists(
    f'{data_dir}pixel_data_agg_TRAIN.csv',
    lambda: group_and_aggregate(train_data)
)

write_csv_from_df(train_data_agg, f'{data_dir}pixel_data_agg_TRAIN')
print(train_data_agg.shape)
train_data_agg.head()

### Test

test_data_agg = get_df_from_csv_if_exists(
    f'{data_dir}pixel_data_agg_TEST.csv',
    lambda: group_and_aggregate(test_data)
)

write_csv_from_df(test_data_agg, f'{data_dir}pixel_data_agg_TEST')
print(test_data_agg.shape)
test_data_agg.head()

## Conclusions and Next Steps

In this notebook, we downloaded the AgrifieldNet competition data, read pixel values from each GeoTiff, added vegetation indices, and aggregated our values by field_id storing the results in CSVs along the way.

Next we will perfom EDA to gain some familiarity with the data before modeling.
