==> code/notebooks/3. Modeling.ipynb <==
# Modeling

This notebook contains code that creates and tunes multiple classification models utilizing Random Forest (RF), Scalar Vector Machine (SVM), XGradient boost (XGB), and Multi-layer Perceptrons (MLP) algorithms. The final model is a voting classifier that utilizes each model as an estimator, where each estimator votes on the probability that a given field contains a particular crop. Ideally, this yields an ensemble model with greater predictive power than any single estimator could achieve working solo.

## Setup

Import the libraries + set the seed val

import pickle
import pandas as pd
import seaborn as sns

from sklearn.ensemble import RandomForestClassifier
from sklearn.model_selection import train_test_split, GridSearchCV
from sklearn.preprocessing import LabelEncoder, RobustScaler
from sklearn.feature_selection import VarianceThreshold
from sklearn.ensemble import VotingClassifier, RandomForestClassifier
from xgboost import XGBClassifier
from sklearn import svm
from sklearn.neural_network import MLPClassifier

from imblearn.over_sampling import SMOTE
from imblearn.pipeline import Pipeline

import sys
sys.path.append('..')

from utils.file_ops import *
from utils.runtime_helpers import *

seed_val = 922

# get config
CONFIG = read_yaml('../../conf.yaml')

# get crops for crop_id -> crop_name resolution
crops = CONFIG['CROPS']

Load the data and add Crop Labels to Train Data

data_dir = '../../data/'

train_labels = pd.read_csv(f'{data_dir}labels_TRAIN.csv', index_col=[0])
train_data_agg = pd.read_csv(f'{data_dir}pixel_data_agg_TRAIN.csv', index_col=[0])
test_data_agg = pd.read_csv(f'{data_dir}pixel_data_agg_TEST.csv', index_col=[0])

train_data_and_labels = train_data_agg.merge(train_labels, on=['field_id'])
print(train_data_and_labels.shape)
train_data_and_labels.head()

## Create the Train / Test Splits

While the Radiant MLHub's AgriFieldNet Competition Dataset ships with test data, the test dataset does not contain crop ids, our predictor (y) values. In the following code block, we split the training data into Test/Train splits so we may assess (validate) the effectiveness of our models on unseen data.

We know this dataset is highly unbalanced, so we will utilize stratification to ensure each split contains data that matches the origin data's distribution of crop types. 

X = train_data_and_labels.drop(['field_id', 'crop_id'], axis=1)
y = train_data_and_labels['crop_id']

# split
X_train, X_test, y_train, y_test = train_test_split(
    X, y,
    test_size=0.33,
    random_state=seed_val,
    stratify=y
)

# save the collection of field ids
field_ids = test_data_agg['field_id']

# encode the labels (needed by XGB)
le = LabelEncoder()
le.fit(y_train)
y_train = le.transform(y_train)
y_test = le.transform(y_test)

print(X_train.shape)
print(X_test.shape)

## Baseline Model

Radiant Earth includes a [starter notebook](https://github.com/radiantearth/agrifieldnet_india_competition/blob/main/Starter%20notebook.ipynb) that contains a baseline model for the AgrifieldNet India Competition. We can recreate and score that model using the following code.


baseline_agg_idxs = ['B01', 'B02', 'B03', 'B04', 'B05', 'B06', 'B07', 'B08', 'B8A', 'B09', 'B11', 'B12']
baseline_agg_metrics = [ '_mean' ]
baseline_selected = get_features(baseline_agg_idxs, baseline_agg_metrics)

baseline_model = RandomForestClassifier(
    n_estimators=20, 
    random_state=0, 
    n_jobs=3)
baseline_model.fit(X_train[baseline_selected], y_train)

(
    baseline_model_y_prob, baseline_model_y_pred, baseline_model_acc, baseline_model_auc, baseline_model_ll
) = eval_and_print_metrics( baseline_model, X_train[baseline_selected], y_train, X_test[baseline_selected], y_test )

The results for our baseline model agree with Radiant Earth's baseline model acheiving approximately 61% accuracy, 74% AUC, and a logloss of 2.57. Note that the logloss calculated above agrees with the competition benchmark found on the [leaderboard](https://zindi.africa/competitions/agrifieldnet-india-challenge/leaderboard) which is 2.564949357.

## Modeling

We will use pipelines to preprocess and model our data. Each model will utilize identical preprocessing steps composed of scaling, feature selection, and oversampling of minority classes. 

- scaling: given that our features contain potential outliers, we utilize the `RobustScaler` as it scales features using statistics robust to outliers.
- programmatic feature selection: `VarianceThreshold` removes low-variance features to improve the predictive power of our data (theoretically, at least.)
- oversampling: crop types are not equally represented (imbalanced). Using `SMOTE`, we will oversample the minority classes in the data to attempt to overcome this.

*Note: Past iterations also used PCA in the preprocessing pipeline, but we determined that PCA reduced our models' predictive power as implemented.*

Below we utilize GridSearch to perform hyperparameter tuning for each model. We are using log loss as our scoring metric as it is better suited for evaluation than the default scoring metric, accuracy, due to the imbalanced nature of our data. We will Then, we execute and save a (somewhat tuned) model.

### Random Forest


param_grid = {
    'classifier__n_estimators': [350, 400, 500],
    'classifier__max_depth': [30, 40, 50],
}

rf_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', RandomForestClassifier(random_state=seed_val))
])

rf_p_cv = GridSearchCV(
    estimator=rf_p,
    param_grid=param_grid, 
    scoring='neg_log_loss', 
    cv=4,
    verbose=False
).fit(X_train, y_train)

print(rf_p_cv.best_score_)
print(rf_p_cv.best_params_)

rf_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', RandomForestClassifier(
        random_state=seed_val,
        max_depth=40,
        n_estimators=500
        ))
])
rf_p.fit(X_train, y_train)

(
    rf_p_y_prob, rf_p_y_pred, rf_p_acc, rf_p_auc, rf_p_ll
) = eval_and_print_metrics( rf_p, X_train, y_train, X_test, y_test )

filename='../../saved_models/rf_p.sav'
pickle.dump(rf_p, open(filename, 'wb'))

### X Gradient Boost


param_grid = {
    'classifier__n_estimators': [100, 120, 150],
    'classifier__max_depth': [10, 15, 20],
    'classifier__learning_rate': [0.09, 0.1, 0.2]
}

xgb_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', XGBClassifier(random_state=seed_val))
])

xgb_p_cv = GridSearchCV(
    estimator=xgb_p,
    param_grid=param_grid, 
    scoring='neg_log_loss',
    cv=4,
    verbose=False
).fit(X_train, y_train)

print(xgb_p_cv.best_score_)
print(xgb_p_cv.best_params_)

xgb_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', XGBClassifier(
        random_state=seed_val,
        learning_rate=0.09,
        max_depth=10,
        n_estimators=100))
])
xgb_p.fit(X_train, y_train)


(
    xgb_p_y_prob, xgb_p_y_pred, xgb_p_acc, xgb_p_auc, xgb_p_ll
) = eval_and_print_metrics( xgb_p, X_train, y_train, X_test, y_test )

filename='../../saved_models/xgb_p.sav'
pickle.dump(xgb_p, open(filename, 'wb'))

### Scalable Vector Machines

param_grid = {
    'classifier__kernel': ['poly', 'rbf', 'sigmoid'],
    'classifier__decision_function_shape': ['ovr', 'ovo'],
}

svm_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', svm.SVC(
        random_state=seed_val,
        probability=True))
])

svm_p_cv = GridSearchCV(
    estimator=svm_p,
    param_grid=param_grid, 
    scoring='neg_log_loss', 
    cv=4,
    verbose=False
).fit(X_train, y_train)

print(svm_p_cv.best_score_)
print(svm_p_cv.best_params_)

svm_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', svm.SVC(
        random_state=seed_val,
        probability=True,
        kernel='rbf',
        decision_function_shape='ovr',
        ))
])

svm_p.fit(X_train, y_train)

(
    svm_p_y_prob, svm_p_y_pred, svm_p_acc, svm_p_auc, svm_p_ll
) = eval_and_print_metrics( svm_p, X_train, y_train, X_test, y_test )

filename='../../saved_models/svm_p.sav'
pickle.dump(svm_p, open(filename, 'wb'))

### Multi-Layer Perceptron

np.arange(start=0.00008, stop=0.00011, step=0.00001)

param_grid = {
    'classifier__hidden_layer_sizes': [8, 16, 32, 64],
    'classifier__solver': ['sgd', 'adam'],
    'classifier__alpha': np.arange(start=0.00008, stop=0.0005, step=0.0001),
}

mlp_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', MLPClassifier(
        random_state=seed_val,
        learning_rate='constant',
        max_iter=4000))
])

mlp_p_cv = GridSearchCV(
    estimator=mlp_p,
    param_grid=param_grid, 
    scoring='neg_log_loss', 
    cv=4,
    verbose=False
).fit(X_train, y_train)

print(mlp_p_cv.best_score_)
print(mlp_p_cv.best_params_)

mlp_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', MLPClassifier(
        random_state=seed_val,
        learning_rate='constant',
        max_iter=4000,
        alpha=8e-05,
        hidden_layer_sizes=8,
        solver='sgd'
        ))
])
# {'alpha': 0.00038, 'hidden_layer_sizes': 32, 'learning_rate': 'constant', 'max_iter': 4000, 'solver': 'adam'}
mlp_p.fit(X_train, y_train)

(
    mlp_p_y_prob, mlp_p_y_pred, mlp_p_acc, mlp_p_auc, mlp_p_ll
) = eval_and_print_metrics( mlp_p, X_train, y_train, X_test, y_test )

filename='../../saved_models/mlp_p.sav'
pickle.dump(mlp_p, open(filename, 'wb'))

### Putting it all together (Voting Classifier)

estimators = [
    ('rf', RandomForestClassifier(
            random_state=seed_val,
            max_depth=40,
            n_estimators=500
        )),
    ('xgb', XGBClassifier(
            random_state=seed_val,
            learning_rate=0.09, 
            max_depth=10, 
            n_estimators=100
    )),
    ('svm', svm.SVC(
            random_state=seed_val,
            decision_function_shape='ovr',
            kernel='rbf',
            probability=True
        )),
    ('mlp', MLPClassifier(
            random_state=seed_val,
            learning_rate='constant', 
            max_iter= 4000, 
            alpha=8e-05,
            hidden_layer_sizes=8,
            solver='sgd'
        ))
]

vc_p = Pipeline([
    ('scaler', RobustScaler()),
    ('selector', VarianceThreshold(threshold=0.001)),
    ('smote', SMOTE(random_state=seed_val)),  
    ('classifier', VotingClassifier(estimators=estimators, voting='soft'))
])
vc_p.fit(X_train, y_train)

(
    vc_p_y_prob, vc_p_y_pred, vc_p_acc, vc_p_auc, vc_p_ll
) = eval_and_print_metrics( vc_p, X_train, y_train, X_test, y_test )

filename = '../../saved_models/vc_p.sav'
pickle.dump(vc_p, open(filename, 'wb'))

## Conclusions and Next Steps

Above we created, tuned, and evaluated models utilizing Random Forest (RF), Scalar Vector Machine (SVM), XGradient boost (XGB), and Multi-layer Perceptrons (MLP) algorithms. The XGB and RF models performed the best solo. But we opted to then combine the predictive power of all 4 models using a voting classifier.

Looking at the classification report for each model, we see that some classes are easier to classify than others. 


crop_names = [get_crop_name_from_id(crops, cid) for cid in le.classes_]
le_resolution = pd.DataFrame(crop_names, columns=['crop_name'])
le_resolution.index.name = 'label'
le_resolution

Using a the above resolution map alongside the classification reports, we found that some crops were more straightforward to classify than others. Well-represented crops, wheat, mustard, maize, and no crop/fallow, were predicted with relatively good precision. In contrast, underrepresented crops, lentils, green peas, garlic, coriander, potatoes, and bersem, were classified with lower precision. Even though we oversampled our minority classes, there were likely still too fields containing these crops in the provided samples to effectively create _valuable_ synthetic data. 

In the next notebook we will refit our models, obtain predictions, and prepare the results for submission to the AgrifieldNet competition site.
