==> code/notebooks/4. Score Submission.ipynb <==
# Creating a submission file

## Setup

Import the libraries + set the seed val

import pickle
import pandas as pd

from sklearn.preprocessing import LabelEncoder

import sys
sys.path.append('..')

from utils.file_ops import *
from utils.runtime_helpers import *
from utils.data_prep import *

seed_val = 922

Load the data and add Crop Labels to Train Data

train_labels = pd.read_csv('../../data/labels_TRAIN.csv', index_col=[0])
train_data_agg = pd.read_csv('../../data/pixel_data_agg_TRAIN.csv', index_col=[0])
test_data_agg = pd.read_csv('../../data/pixel_data_agg_TEST.csv', index_col=[0])

train_data_and_labels = train_data_agg.merge(train_labels, on=['field_id'])

## Assign Train / Test values

Below we will assign our X and y values using the complete train dataset and the test dataset - unlike our prior modeling efforts which split the train set into train/test collections.

X_train = train_data_and_labels.drop(['field_id', 'crop_id'], axis=1)
y_train = train_data_and_labels['crop_id']
X_test = test_data_agg.drop(['field_id'], axis=1)
field_ids = test_data_agg['field_id']

le = LabelEncoder()
le.fit(y_train)
y_train = le.transform(y_train)

Additionally, experimentation has shown that using the following values for modeling (obtained using our PCA in the EDA notebook) appears to lower the logloss and improve the score for the final submission

*Note: these same selected values were also used in prior iterations for the modeling notebook that preceeds this one, however, we obtained better metrics using all the features available to us in that context.

agg_idxs = [
    'B01', 'B02', 'B03', 'B04', 'B05', 'B06', 'B07', 'B08', 'B8A', 'B09', 'B11', 'B12',
    'NDVI', 'ARVI', 'SAVI', 'MSI', 'MCARI', 'MARI', 'EVI2', 'NDMI', 'NDWI',
    'brightness'
]    
agg_metrics = [
    '_median'
]

selected = get_features(agg_idxs, agg_metrics, ['pixels'])

## Fitting / Preparing Predictions for Submission

### Loading the saved models

rf_p = pickle.load(open('../../saved_models/rf_p.sav', 'rb'))
xgb_p = pickle.load(open('../../saved_models/xgb_p.sav', 'rb'))
svm_p = pickle.load(open('../../saved_models/svm_p.sav', 'rb'))
mlp_p = pickle.load(open('../../saved_models/mlp_p.sav', 'rb'))
vp_p = pickle.load(open('../../saved_models/vc_p.sav', 'rb'))


### Random Forest

clf = rf_p

clf.fit(X_train[selected], y_train)
y_test_pred = clf.predict_proba(X_test[selected])

crop_dict = get_crop_dict()
crop_columns = [crop_dict.get(i) for i in le.inverse_transform(clf.classes_)]

test_df = pd.DataFrame(columns= ['field_id'] + crop_columns)
test_df['field_id'] = field_ids
test_df[crop_columns] = y_test_pred 

test_df.to_csv('../../submissions/rf_p.csv', index=False)

test_df.head()

### X Gradient Boost

clf = xgb_p

clf.fit(X_train[selected], y_train)
y_test_pred = clf.predict_proba(X_test[selected])

crop_dict = get_crop_dict()
crop_columns = [crop_dict.get(i) for i in le.inverse_transform(clf.classes_)]

test_df = pd.DataFrame(columns= ['field_id'] + crop_columns)
test_df['field_id'] = field_ids
test_df[crop_columns] = y_test_pred 

test_df.to_csv('../../submissions/xgb_p.csv', index=False)

test_df.head()

1.758395241
1.765851258

### Scalable Vector Machines

clf = svm_p

clf.fit(X_train[selected], y_train)
y_test_pred = clf.predict_proba(X_test[selected])

crop_dict = get_crop_dict()
crop_columns = [crop_dict.get(i) for i in le.inverse_transform(clf.classes_)]

test_df = pd.DataFrame(columns= ['field_id'] + crop_columns)
test_df['field_id'] = field_ids
test_df[crop_columns] = y_test_pred 

test_df.to_csv('../../submissions/svm_p.csv', index=False)

test_df.head()

### Multi-Layer Perceptron

clf = mlp_p

clf.fit(X_train[selected], y_train)
y_test_pred = clf.predict_proba(X_test[selected])

crop_dict = get_crop_dict()
crop_columns = [crop_dict.get(i) for i in le.inverse_transform(clf.classes_)]

test_df = pd.DataFrame(columns= ['field_id'] + crop_columns)
test_df['field_id'] = field_ids
test_df[crop_columns] = y_test_pred 

test_df.to_csv('../../submissions/mlp_p.csv', index=False)

test_df.head()

# Voting Classifier

clf = vp_p

clf.fit(X_train[selected], y_train)
y_test_pred = clf.predict_proba(X_test[selected])

crop_dict = get_crop_dict()
crop_columns = [crop_dict.get(i) for i in le.inverse_transform(clf.classes_)]

test_df = pd.DataFrame(columns= ['field_id'] + crop_columns)
test_df['field_id'] = field_ids
test_df[crop_columns] = y_test_pred 

test_df.to_csv('../../submissions/vp_p.csv', index=False)

test_df.head()

## Analysis

After submitting each model's predictions on the test data, we achieve the following scores from the AgrifieldNet competition.

| Model | Public Score | Private Score |
| ----- | ------------ | ------------- |
| rf_p  | 1.744836782  | 2.163112415   |
| xgb_p | 1.758395241  | 1.765851258   |
| svm_p | 1.465262957  | 1.497996049   |
| mlp_p | 1.806023962  | 1.767157053   |
| v_p   | 1.309915641  | 1.372958035   |

The competition website explains the presence of both public and private scoring by stating that "Zindi maintains a public leaderboard and a private leaderboard for each competition. The Public Leaderboard includes approximately 20% of the test dataset."

As hypothesized, the voting classifier performs better than any other estimators we created when evaluated on their own. While we wouldn't have won the competition with our voting classifier model, it does put us at **22nd** place out of 151 submissions according to the [leaderboard](https://zindi.africa/competitions/agrifieldnet-india-challenge/leaderboard) (had we submitted our results before the end of the cutoff date.)

## Conclusions and Possible Next Steps

We fit and scored models we created in the previous notebook. We then processed and wrote the results into CSV files for submission to the AgrifieldNet competition website.

Interestingly, according to the AgrifieldNet scores, our models performed better with a subset of features. However, our models benefitted from all available features when evaluated locally using training data split into train/test splits. Further exploration into the effects of manual and automated feature selection on the predictive power of our models is warranted. The voting classifier could also benefit from hyperparameter tuning, which was not performed due to a lack of time/compute power.
